requirements.txt
Django
gunicorn
mysqlclient
python-dotenv
django-extensions
django-template-partials
django-htmx
django-crispy-forms
crispy-bootstrap5